Added a GameState object which can be used by all other objects as a way to globally track game state variables.
This will allow objects to reference certain game state information as needed, and it will allow Events to subscribe to changes in global game state variables to trigger at the appropriate time. This will bypass a need to constantly update events periodically that would have to poll other objects for changes.