Added some utility functions for doing things like vector computations.
Adjusted some GameObject methods to use the new Util methods and also changed all dx/dy parameters to use CGVector instead with improved code readability.
Enemies now fire to the direction of the player. This is currently wrong because of the way the GameScene coordinates are set up, which needs to be modified.
Also added the option to set bullet colors.