Fixed how positions are tracked through GameObjects.
A new getPosition() method must be used, which automatically decides if the node exists or if returning the default position.
Also added a new Util.getOrientation(from: CGVector) method that handles returning a rotation from a directional vector.
TODO: Bullets are re-oriented using this new getOrientation method, but it doesn't work correctly if the direction vector is negative in the X direction.