CreateExperienceToken is now a general CreateLootPackage class.
This creates a loot item which contains experience for now, and other things can be added too.
The purpose of loot items is to be more dramatic, and collide with the player. So far, the player has to move over to the loot item. The next step is to make LootItem move and potentially come towards the player.