Added an ugly background to the scene.
All nodes are now added to this background node. The idea is that the background node is larger than the scene, so the world can be larger than is visible directly on the screen. It will also make scaling more straight-forward since the world will be square. That additional functionality needs to be worked out.